+++ src/compiler/core/src/commonMain/kotlin/community/flock/wirespec/compiler/core/emit/HasPackageName.kt
@@ -1,5 +1,5 @@
 package community.flock.wirespec.compiler.core.emit
 
-interface PackageNameEmitter {
+interface HasPackageName {
     val packageName: PackageName
 }
