+++ src/compiler/core/src/commonMain/kotlin/community/flock/wirespec/compiler/core/parse/Node.kt
@@ -188,8 +188,14 @@ value class Comment private constructor(override val value: String) : Value<Stri
 sealed class Identifier(name: String) : Value<String> {
     override val value = name.removeBackticks()
     override fun toString() = value
+
+    abstract fun map(fn: (String) -> String): Identifier
 }
 
-data class DefinitionIdentifier(private val name: String) : Identifier(name)
+data class DefinitionIdentifier(private val name: String) : Identifier(name) {
+    override fun map(fn: (String) -> String) = DefinitionIdentifier(fn(value))
+}
 
-data class FieldIdentifier(private val name: String) : Identifier(name)
+data class FieldIdentifier(private val name: String) : Identifier(name) {
+    override fun map(fn: (String) -> String) = FieldIdentifier(fn(value))
+}
