+++ src/compiler/emitters/java/src/commonMain/kotlin/community/flock/wirespec/emitters/java/JavaChannelDefinitionEmitter.kt
@@ -1,14 +1,13 @@
 package community.flock.wirespec.emitters.java
 
 import community.flock.wirespec.compiler.core.emit.ChannelDefinitionEmitter
-import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
-import community.flock.wirespec.compiler.core.emit.ImportEmitter
-import community.flock.wirespec.compiler.core.emit.PackageNameEmitter
+import community.flock.wirespec.compiler.core.emit.HasPackageName
+import community.flock.wirespec.compiler.core.emit.importReferences
 import community.flock.wirespec.compiler.core.parse.Channel
 import community.flock.wirespec.compiler.core.parse.Definition
 import community.flock.wirespec.compiler.core.parse.Reference
 
-interface JavaChannelDefinitionEmitter: JavaTypeDefinitionEmitter, PackageNameEmitter, ChannelDefinitionEmitter, ImportEmitter, IdentifierEmitter {
+interface JavaChannelDefinitionEmitter: ChannelDefinitionEmitter, HasPackageName, JavaTypeDefinitionEmitter  {
 
     override fun emit(channel: Channel) = """
         |${channel.emitImports()}
