+++ src/compiler/emitters/java/src/commonMain/kotlin/community/flock/wirespec/emitters/java/JavaIdentifierEmitter.kt
@@ -2,10 +2,10 @@ package community.flock.wirespec.emitters.java
 
 import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.firstToUpper
 import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
+import community.flock.wirespec.compiler.core.emit.Keywords
 import community.flock.wirespec.compiler.core.parse.DefinitionIdentifier
 import community.flock.wirespec.compiler.core.parse.FieldIdentifier
 import community.flock.wirespec.compiler.core.parse.Identifier
-import community.flock.wirespec.emitters.java.JavaEmitter.Companion.reservedKeywords
 
 interface JavaIdentifierEmitter: IdentifierEmitter {
 
@@ -27,4 +27,19 @@ interface JavaIdentifierEmitter: IdentifierEmitter {
 
     fun String.sanitizeKeywords() = if (this in reservedKeywords) "_$this" else this
 
+    companion object : Keywords {
+        override val reservedKeywords = setOf(
+            "abstract", "continue", "for", "new", "switch",
+            "assert", "default", "goto", "package", "synchronized",
+            "boolean", "do", "if", "private", "this",
+            "break", "double", "implements", "protected", "throw",
+            "byte", "else", "import", "public", "throws",
+            "case", "enum", "instanceof", "return", "transient",
+            "catch", "extends", "int", "short", "try",
+            "char", "final", "interface", "static", "void",
+            "class", "finally", "long", "strictfp", "volatile",
+            "const", "float", "native", "super", "while",
+            "true", "false"
+        )
+    }
 }
