+++ src/compiler/emitters/kotlin/src/commonMain/kotlin/community/flock/wirespec/emitters/kotlin/KotlinIdentifierEmitter.kt
@@ -1,21 +1,21 @@
 package community.flock.wirespec.emitters.kotlin
 
 import community.flock.wirespec.compiler.core.addBackticks
-import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.firstToUpper
 import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
+import community.flock.wirespec.compiler.core.emit.Keywords
+import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.firstToUpper
 import community.flock.wirespec.compiler.core.parse.DefinitionIdentifier
 import community.flock.wirespec.compiler.core.parse.FieldIdentifier
 import community.flock.wirespec.compiler.core.parse.Identifier
-import community.flock.wirespec.emitters.kotlin.KotlinEmitter.Companion.reservedKeywords
 
-interface KotlinIdentifierEmitter: IdentifierEmitter {
+interface KotlinIdentifierEmitter : IdentifierEmitter {
 
     override fun emit(identifier: Identifier) = when (identifier) {
         is DefinitionIdentifier -> identifier.sanitize()
         is FieldIdentifier -> identifier.sanitize().sanitizeKeywords()
     }
 
-    override fun Identifier.sanitize() = this.value
+  fun Identifier.sanitize() = value
         .split(".", " ")
         .mapIndexed { index, s -> if (index > 0) s.firstToUpper() else s }
         .joinToString("")
@@ -28,4 +28,14 @@ interface KotlinIdentifierEmitter: IdentifierEmitter {
 
     fun String.sanitizeKeywords() = if (this in reservedKeywords) addBackticks() else this
 
+    companion object : Keywords {
+        override val reservedKeywords = setOf(
+            "as", "break", "class", "continue", "do",
+            "else", "false", "for", "fun", "if",
+            "in", "interface", "internal", "is", "null",
+            "object", "open", "package", "return", "super",
+            "this", "throw", "true", "try", "typealias",
+            "typeof", "val", "var", "when", "while", "private", "public"
+        )
+    }
 }
