+++ src/compiler/emitters/kotlin/src/commonMain/kotlin/community/flock/wirespec/emitters/kotlin/KotlinRefinedTypeDefinitionEmitter.kt
@@ -1,13 +1,11 @@
 package community.flock.wirespec.emitters.kotlin
 
-import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
 import community.flock.wirespec.compiler.core.emit.RefinedTypeDefinitionEmitter
 import community.flock.wirespec.compiler.core.emit.Spacer
-import community.flock.wirespec.compiler.core.emit.TypeDefinitionEmitter
 import community.flock.wirespec.compiler.core.parse.Reference
 import community.flock.wirespec.compiler.core.parse.Refined
 
-interface KotlinRefinedTypeDefinitionEmitter: RefinedTypeDefinitionEmitter, IdentifierEmitter, TypeDefinitionEmitter {
+interface KotlinRefinedTypeDefinitionEmitter: RefinedTypeDefinitionEmitter, KotlinIdentifierEmitter, KotlinTypeDefinitionEmitter {
 
     override fun emit(refined: Refined) = """
         |data class ${refined.identifier.sanitize()}(override val value: String): Wirespec.Refined {
@@ -18,6 +16,15 @@ interface KotlinRefinedTypeDefinitionEmitter: RefinedTypeDefinitionEmitter, Iden
         |
     """.trimMargin()
 
-
+    override fun Refined.emitValidator():String {
+        val defaultReturn = "true"
+        return when (val type = reference.type) {
+            is Reference.Primitive.Type.Integer -> type.constraint?.emit() ?: defaultReturn
+            is Reference.Primitive.Type.Number -> type.constraint?.emit() ?: defaultReturn
+            is Reference.Primitive.Type.String -> type.constraint?.emit() ?: defaultReturn
+            Reference.Primitive.Type.Boolean -> defaultReturn
+            Reference.Primitive.Type.Bytes -> defaultReturn
+        }
+    }
 
 }
