+++ src/compiler/emitters/kotlin/src/commonMain/kotlin/community/flock/wirespec/emitters/kotlin/KotlinTypeDefinitionEmitter.kt
@@ -1,15 +1,13 @@
 package community.flock.wirespec.emitters.kotlin
 
-import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
 import community.flock.wirespec.compiler.core.emit.Spacer
 import community.flock.wirespec.compiler.core.emit.TypeDefinitionEmitter
 import community.flock.wirespec.compiler.core.parse.Field
 import community.flock.wirespec.compiler.core.parse.Module
 import community.flock.wirespec.compiler.core.parse.Reference
-import community.flock.wirespec.compiler.core.parse.Refined
 import community.flock.wirespec.compiler.core.parse.Type
 
-interface KotlinTypeDefinitionEmitter: TypeDefinitionEmitter, IdentifierEmitter {
+interface KotlinTypeDefinitionEmitter : TypeDefinitionEmitter, KotlinIdentifierEmitter {
 
     override fun emit(type: Type, module: Module) =
         if (type.shape.value.isEmpty()) "data object ${emit(type.identifier)}"
@@ -47,18 +45,7 @@ interface KotlinTypeDefinitionEmitter: TypeDefinitionEmitter, IdentifierEmitter 
         }
     }.let { if (isNullable) "$it?" else it }
 
-    override fun Refined.emitValidator():String {
-        val defaultReturn = "true"
-        return when (val type = reference.type) {
-            is Reference.Primitive.Type.Integer -> type.constraint?.emit() ?: defaultReturn
-            is Reference.Primitive.Type.Number -> type.constraint?.emit() ?: defaultReturn
-            is Reference.Primitive.Type.String -> type.constraint?.emit() ?: defaultReturn
-            Reference.Primitive.Type.Boolean -> defaultReturn
-            Reference.Primitive.Type.Bytes -> defaultReturn
-        }
-    }
-
-    override fun Reference.Primitive.Type.Constraint.emit() = when(this){
+    override fun Reference.Primitive.Type.Constraint.emit() = when (this) {
         is Reference.Primitive.Type.Constraint.RegExp -> "Regex(\"\"\"$expression\"\"\").matches(value)"
         is Reference.Primitive.Type.Constraint.Bound -> """${Spacer}$min < record.value < $max;"""
     }
