+++ src/compiler/emitters/python/src/commonMain/kotlin/community/flock/wirespec/emitters/python/PythonChannelDefinitionEmitter.kt
@@ -3,8 +3,6 @@ package community.flock.wirespec.emitters.python
 import community.flock.wirespec.compiler.core.emit.ChannelDefinitionEmitter
 import community.flock.wirespec.compiler.core.parse.Channel
 
-interface PythonChannelDefinitionEmitter: ChannelDefinitionEmitter {
-
+interface PythonChannelDefinitionEmitter : ChannelDefinitionEmitter {
     override fun emit(channel: Channel) = ""
-
 }
