+++ src/compiler/emitters/typescript/src/commonMain/kotlin/community/flock/wirespec/emitters/typescript/TypeScriptRefinedTypeDefinitionEmitter.kt
@@ -2,6 +2,7 @@ package community.flock.wirespec.emitters.typescript
 
 import community.flock.wirespec.compiler.core.emit.RefinedTypeDefinitionEmitter
 import community.flock.wirespec.compiler.core.emit.Spacer
+import community.flock.wirespec.compiler.core.parse.Reference
 import community.flock.wirespec.compiler.core.parse.Refined
 
 interface TypeScriptRefinedTypeDefinitionEmitter: RefinedTypeDefinitionEmitter, TypeScriptTypeDefinitionEmitter {
@@ -12,4 +13,15 @@ interface TypeScriptRefinedTypeDefinitionEmitter: RefinedTypeDefinitionEmitter, 
           |${Spacer}${refined.emitValidator()};
           |
         """.trimMargin()
+
+    override fun Refined.emitValidator(): String {
+        val defaultReturn = "true;"
+        return when (val type = reference.type) {
+            is Reference.Primitive.Type.Integer -> type.constraint?.emit() ?: defaultReturn
+            is Reference.Primitive.Type.Number -> type.constraint?.emit() ?: defaultReturn
+            is Reference.Primitive.Type.String -> type.constraint?.emit() ?: defaultReturn
+            Reference.Primitive.Type.Boolean -> defaultReturn
+            Reference.Primitive.Type.Bytes -> defaultReturn
+        }
+    }
 }
