+++ src/compiler/emitters/wirespec/src/commonMain/kotlin/community/flock/wirespec/emitters/wirespec/WirespecChannelDefinitionEmitter.kt
@@ -5,9 +5,7 @@ import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
 import community.flock.wirespec.compiler.core.emit.TypeDefinitionEmitter
 import community.flock.wirespec.compiler.core.parse.Channel
 
-interface WirespecChannelDefinitionEmitter: ChannelDefinitionEmitter, IdentifierEmitter, TypeDefinitionEmitter {
-
+interface WirespecChannelDefinitionEmitter: ChannelDefinitionEmitter, WirespecTypeDefinitionEmitter {
     override fun emit(channel: Channel): String =
         "channel ${emit(channel.identifier)} -> ${channel.reference.emit()}"
-
 }
