+++ src/compiler/emitters/wirespec/src/commonMain/kotlin/community/flock/wirespec/emitters/wirespec/WirespecEmitter.kt
@@ -5,7 +5,7 @@ import community.flock.wirespec.compiler.core.emit.FileExtension
 import community.flock.wirespec.compiler.core.emit.Keywords
 import community.flock.wirespec.compiler.core.parse.Reference
 
-interface E:
+interface WirespecEmitters:
     WirespecIdentifierEmitter,
     WirespecTypeDefinitionEmitter,
     WirespecEndpointDefinitionEmitter,
@@ -14,7 +14,7 @@ interface E:
     WirespecUnionDefinitionEmitter,
     WirespecRefinedTypeDefinitionEmitter
 
-open class WirespecEmitter : LanguageEmitter(), E {
+open class WirespecEmitter : LanguageEmitter(), WirespecEmitters {
 
     override val extension = FileExtension.Wirespec
 
@@ -27,11 +27,5 @@ open class WirespecEmitter : LanguageEmitter(), E {
     override fun Reference.Primitive.Type.Constraint.emit() = when(this){
         is Reference.Primitive.Type.Constraint.RegExp -> "(${value})"
         is Reference.Primitive.Type.Constraint.Bound -> "(${min}, ${max})"
-    }
-
-    companion object : Keywords {
-        override val reservedKeywords = setOf(
-            "type", "enum", "endpoint"
-        )
     }
 }
