+++ src/compiler/emitters/wirespec/src/commonMain/kotlin/community/flock/wirespec/emitters/wirespec/WirespecEnumDefinitionEmitter.kt
@@ -6,7 +6,7 @@ import community.flock.wirespec.compiler.core.emit.Spacer
 import community.flock.wirespec.compiler.core.parse.Enum
 import community.flock.wirespec.compiler.core.parse.Module
 
-interface WirespecEnumDefinitionEmitter: EnumDefinitionEmitter, IdentifierEmitter {
+interface WirespecEnumDefinitionEmitter: EnumDefinitionEmitter, WirespecIdentifierEmitter {
 
     override fun emit(enum: Enum, module: Module) =
         "enum ${emit(enum.identifier)} {\n${Spacer}${enum.entries.joinToString(", ") { it.capitalize() }}\n}\n"
