+++ src/compiler/emitters/wirespec/src/commonMain/kotlin/community/flock/wirespec/emitters/wirespec/WirespecTypeDefinitionEmitter.kt
@@ -1,15 +1,13 @@
 package community.flock.wirespec.emitters.wirespec
 
-import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
 import community.flock.wirespec.compiler.core.emit.Spacer
 import community.flock.wirespec.compiler.core.emit.TypeDefinitionEmitter
 import community.flock.wirespec.compiler.core.parse.Field
 import community.flock.wirespec.compiler.core.parse.Module
 import community.flock.wirespec.compiler.core.parse.Reference
-import community.flock.wirespec.compiler.core.parse.Refined
 import community.flock.wirespec.compiler.core.parse.Type
 
-interface WirespecTypeDefinitionEmitter: TypeDefinitionEmitter, IdentifierEmitter {
+interface WirespecTypeDefinitionEmitter : TypeDefinitionEmitter, WirespecIdentifierEmitter {
 
     override fun emit(type: Type, module: Module) = """
         |type ${emit(type.identifier)} {
@@ -42,15 +40,5 @@ interface WirespecTypeDefinitionEmitter: TypeDefinitionEmitter, IdentifierEmitte
             }
         }
     }.let { if (isNullable) "$it?" else it }
-
-    override fun Refined.emitValidator():String {
-        return when (val type = reference.type) {
-            is Reference.Primitive.Type.Integer -> type.constraint?.emit() ?: ""
-            is Reference.Primitive.Type.Number -> type.constraint?.emit() ?: ""
-            is Reference.Primitive.Type.String -> type.constraint?.emit() ?: ""
-            Reference.Primitive.Type.Boolean -> ""
-            Reference.Primitive.Type.Bytes -> ""
-        }
-    }
 
 }
