+++ src/compiler/emitters/wirespec/src/commonMain/kotlin/community/flock/wirespec/emitters/wirespec/WirespecUnionDefinitionEmitter.kt
@@ -1,11 +1,9 @@
 package community.flock.wirespec.emitters.wirespec
 
-import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
-import community.flock.wirespec.compiler.core.emit.TypeDefinitionEmitter
 import community.flock.wirespec.compiler.core.emit.UnionDefinitionEmitter
 import community.flock.wirespec.compiler.core.parse.Union
 
-interface WirespecUnionDefinitionEmitter: UnionDefinitionEmitter, IdentifierEmitter, TypeDefinitionEmitter {
+interface WirespecUnionDefinitionEmitter: UnionDefinitionEmitter, WirespecTypeDefinitionEmitter {
 
     override fun emit(union: Union) =
         "type ${emit(union.identifier)} = ${union.entries.joinToString(" | ") { it.emit() }}\n"
