+++ src/integration/avro/src/jvmMain/kotlin/community/flock/wirespec/integration/avro/kotlin/emit/AvroEmitter.kt
@@ -3,6 +3,7 @@ package community.flock.wirespec.integration.avro.kotlin.emit
 import community.flock.wirespec.compiler.core.emit.EmitShared
 import community.flock.wirespec.compiler.core.emit.PackageName
 import community.flock.wirespec.compiler.core.emit.Spacer
+import community.flock.wirespec.compiler.core.emit.spacer
 import community.flock.wirespec.compiler.core.parse.Definition
 import community.flock.wirespec.compiler.core.parse.Enum
 import community.flock.wirespec.compiler.core.parse.Field
