+++ src/integration/spring/src/jvmMain/kotlin/community/flock/wirespec/integration/spring/kotlin/emit/SpringKotlinEmitter.kt
@@ -3,6 +3,7 @@ package community.flock.wirespec.integration.spring.kotlin.emit
 import community.flock.wirespec.compiler.core.emit.EmitShared
 import community.flock.wirespec.compiler.core.emit.PackageName
 import community.flock.wirespec.compiler.core.emit.Spacer
+import community.flock.wirespec.compiler.core.emit.emit
 import community.flock.wirespec.compiler.core.parse.Endpoint
 import community.flock.wirespec.emitters.kotlin.KotlinEmitter
